Move __getattr__ back to __init__.py
